Move train files to sleap_nn
